Make Syntax more friendly to current Intellij
As of now, Intellij does not take into account matching implicits when suggesting extension methods during autocomplete,
although it did so before, and ostensibly it should still do it. Because of that the autocomplete list for dynamo queries
contains a lot of methods that aren't applicable to the current query (e.g. `withUpdateExpression` for `QueryRequest`),
this makes navigating the library very hard. Here I'm changing most (excluding `decodeItems` & `decodeItem` because of complex overloading)
implicit classes to match on upper bound of DynamoRequest instead of an implicit evidence, this works around the Intellij issues
and massively cleans up the auto-complete list